requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
dj-database-url==1.3.0
Django==4.1.7
gunicorn==20.1.0
Pillow==9.4.0
soupsieve==2.4
sqlparse==0.4.3
typing_extensions==4.5.0
